Renamed SwapStepper component
