+++ common/v2/features/SwapAssets/components/ConfirmSwapMultiTx.stories.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import SwapStepper from './SwapStepper';
+import ConfirmSwapMultiTx from './ConfirmSwapMultiTx';
 import { ISwapAsset } from '../types';
 import { TSymbol } from 'v2/types';
 
-export default { title: 'SwapStepper' };
+export default { title: 'ConfirmSwapMultiTx' };
 
 const DAI: ISwapAsset = { name: 'DAI Stablecoin v2.0', symbol: 'DAI' as TSymbol };
 const ETH: ISwapAsset = { name: 'Ethereum', symbol: 'ETH' as TSymbol };
@@ -11,7 +11,7 @@ const daiAmount = '100';
 const ethAmount = '0.5';
 
 export const daiToEth = () => (
-  <SwapStepper
+  <ConfirmSwapMultiTx
     currentStep={0}
     fromAsset={DAI}
     toAsset={ETH}
@@ -21,7 +21,7 @@ export const daiToEth = () => (
 );
 
 export const daiToEthStep2 = () => (
-  <SwapStepper
+  <ConfirmSwapMultiTx
     currentStep={1}
     fromAsset={DAI}
     toAsset={ETH}
