+++ common/v2/features/SwapAssets/components/index.ts
@@ -2,4 +2,4 @@ export { default as SwapAssets } from './SwapAssets';
 export { default as ConfirmSwap } from './ConfirmSwap';
 export { default as SwapTransactionReceipt } from './SwapTransactionReceipt';
 export { default as SetAllowance } from './SetAllowance';
-export { default as SwapStepper } from './SwapStepper';
+export { default as ConfirmSwapMultiTx } from './ConfirmSwapMultiTx';
